rm redundant files, improve null data values
Electron affinity and eigenvalues have both been using 0.
as a default value. This leads to them being used transparently
when perhaps it would be better to signal an error, and leads to
genuine zero values being indistinguishable from missing data.
The string 'None' is now used in data/element.text to indicate
missing data, and is converted to a Python None object by the data
loader.

This breaks the compound electronegativity test, which uses brass as an
example. It is unable to compute the Mulliken electronegativity of Zn as
the electron affinity is not defined. Admittedly, intermetallics have
other issues in SMACT and so brass may not have been the best choice of
case study...